+++ installer.ps1
@@ -1,12 +1,12 @@
 #!/usr/bin/env pwsh
-# generated with CMTools 2.5.0 45de33a
+# generated with CMTools 2.5.1 29b7fff
 
 #
 # Set the package name and version to install
 #
 param(
   [Parameter()]
-  [String]$VERSION = "2.5.0"
+  [String]$VERSION = "2.5.1"
 )
 [String]$PKG_VERSION = [Environment]::GetEnvironmentVariable("PKG_VERSION")
 if ($PKG_VERSION) {