+++ sqlstore_sqlite.go
@@ -1,0 +1,7 @@
+//go:build !wasip1
+
+package dataset
+
+import _ "github.com/glebarez/go-sqlite"
+
+const Sqlite3DriverName = "sqlite"